requirements.txt
Django
psycopg2
south
requests
pytz
django-timezone-field
djangorestframework
django-celery
